+++ requests_retry_session/timeout_http_adapter.py
@@ -27,6 +27,8 @@ from typing import TYPE_CHECKING
 
 from requests.adapters import HTTPAdapter
 
+from .utils import NotPassed, NOT_PASSED
+
 if TYPE_CHECKING:
     from typing import Mapping, TypedDict, Union
 
@@ -52,7 +54,7 @@ if TYPE_CHECKING:
         cert: CertType
         proxies: ProxiesType
 
-    class _InitArgs(TypedDict, total=False):  # pylint: disable=missing-class-docstring
+    class _InitArgs(TypedDict, total=False):
         """
         The valid kwargs for HTTPAdapter.__init__()
         """
@@ -60,16 +62,6 @@ if TYPE_CHECKING:
         pool_maxsize: int
         max_retries: Union[Retry, int, None]
         pool_block: bool
-
-class _NotPassed:  # pylint: disable=too-few-public-methods
-    """
-    A dummy class to let us distinguish between an argument not being passed versus
-    an argument explicitly being passed with a None value
-    """
-    pass
-
-
-_NOT_PASSED = _NotPassed()
 
 
 class TimeoutHTTPAdapter(HTTPAdapter):
@@ -81,41 +73,41 @@ class TimeoutHTTPAdapter(HTTPAdapter):
 
     def __init__(  # pylint: disable=too-many-arguments,too-many-positional-arguments
             self,
-            pool_connections: Union[int, _NotPassed] = _NOT_PASSED,
-            pool_maxsize: Union[int, _NotPassed] = _NOT_PASSED,
-            max_retries: Union[Retry, int, None, _NotPassed] = _NOT_PASSED,
-            pool_block: Union[bool, _NotPassed] = _NOT_PASSED,
+            pool_connections: Union[int, NotPassed] = NOT_PASSED,
+            pool_maxsize: Union[int, NotPassed] = NOT_PASSED,
+            max_retries: Union[Retry, int, None, NotPassed] = NOT_PASSED,
+            pool_block: Union[bool, NotPassed] = NOT_PASSED,
             timeout: TimeoutType = None) -> None:
         self.timeout: TimeoutType = timeout
         kwargs: _InitArgs = {}
-        if not isinstance(pool_connections, _NotPassed):
+        if not isinstance(pool_connections, NotPassed):
             kwargs["pool_connections"] = pool_connections
-        if not isinstance(pool_maxsize, _NotPassed):
+        if not isinstance(pool_maxsize, NotPassed):
             kwargs["pool_maxsize"] = pool_maxsize
-        if not isinstance(max_retries, _NotPassed):
+        if not isinstance(max_retries, NotPassed):
             kwargs["max_retries"] = max_retries
-        if not isinstance(pool_block, _NotPassed):
+        if not isinstance(pool_block, NotPassed):
             kwargs["pool_block"] = pool_block
         super().__init__(**kwargs)
 
     def send(  # pylint: disable=too-many-arguments,too-many-positional-arguments
             self,
             request: PreparedRequest,
-            stream: Union[bool, _NotPassed] = _NOT_PASSED,
-            timeout: Union[TimeoutType, _NotPassed] = _NOT_PASSED,
-            verify: Union[VerifyType, _NotPassed] = _NOT_PASSED,
-            cert: Union[CertType, _NotPassed] = _NOT_PASSED,
-            proxies: Union[ProxiesType, _NotPassed] = _NOT_PASSED) -> Response:
+            stream: Union[bool, NotPassed] = NOT_PASSED,
+            timeout: Union[TimeoutType, NotPassed] = NOT_PASSED,
+            verify: Union[VerifyType, NotPassed] = NOT_PASSED,
+            cert: Union[CertType, NotPassed] = NOT_PASSED,
+            proxies: Union[ProxiesType, NotPassed] = NOT_PASSED) -> Response:
         kwargs: _SendArgs = {
             "timeout":
-            self.timeout if isinstance(timeout, _NotPassed) else timeout
+            self.timeout if isinstance(timeout, NotPassed) else timeout
         }
-        if not isinstance(stream, _NotPassed):
+        if not isinstance(stream, NotPassed):
             kwargs["stream"] = stream
-        if not isinstance(verify, _NotPassed):
+        if not isinstance(verify, NotPassed):
             kwargs["verify"] = verify
-        if not isinstance(cert, _NotPassed):
+        if not isinstance(cert, NotPassed):
             kwargs["cert"] = cert
-        if not isinstance(proxies, _NotPassed):
+        if not isinstance(proxies, NotPassed):
             kwargs["proxies"] = proxies
         return super().send(request, **kwargs)