+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Removed dependence on Retry.DEFAULT_ALLOWED_METHODS
+
 ## [5.1.0] - 2026-04-15
 
 ### Changed