+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added basic build time tests to catch bugs that break the module in a basic way,
+  for any of the supported Python versions
+
 ## [5.0.3] - 2026-04-15
 
 ### Added