+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added basic build time tests to catch bugs that break the module in a basic way,
   for any of the supported Python versions
 
+### Changed
+- Removed dependence on Retry.DEFAULT_ALLOWED_METHODS
+
 ## [1.1.0] - 2026-04-15
 
 ### Changed